Service, test 기능

+++ src/main/java/com/example/jsontodb/domain/ImageInfo.java
@@ -19,9 +19,9 @@ public class ImageInfo {
     @GeneratedValue
     private Long Id;
 
-    private int height;
+    private Long height;
 
-    private int width;
+    private Long width;
 
     @OneToOne
     @ToString.Exclude
